+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
 
+## [2.22.0] - 2023-04-20
+
 ### Added
 
 - Add head_js block into base html template
@@ -1846,7 +1848,8 @@ us:
 - finish integrating the missing pages and improve the sandbox environment;
 - test and polish the use of richie as a django app / node dependency.
 
-[unreleased]: https://github.com/openfun/richie/compare/v2.21.1...master
+[unreleased]: https://github.com/openfun/richie/compare/v2.22.0...master
+[2.22.0]: https://github.com/openfun/richie/compare/v2.21.1...2.22.0
 [2.21.1]: https://github.com/openfun/richie/compare/v2.21.0...2.21.1
 [2.21.0]: https://github.com/openfun/richie/compare/v2.20.1...2.21.0
 [2.20.1]: https://github.com/openfun/richie/compare/v2.20.0...2.20.1